Move fragment in right package
